+++ util/calc.py
@@ -11,8 +11,9 @@ import ntplib
 import util.stream_request
 from jsonresponse import JsonResponse
 from ooi_data.postgres.model import Stream, Parameter
-from util.common import (StreamKey, TimeRange, MalformedRequestException, InvalidStreamException,
-                         InvalidParameterException, UIHardLimitExceededException, MissingDataException)
+from util.common import (get_annotation_filename, StreamKey, TimeRange, MalformedRequestException, 
+                         InvalidStreamException, InvalidParameterException, UIHardLimitExceededException, 
+                         MissingDataException)
 from util.csvresponse import CsvGenerator
 from util.netcdf_generator import NetcdfGenerator
 from engine import app
@@ -232,8 +233,7 @@ def _write_annotations(stream_request, base_path):
         return
     
     if stream_request.include_annotations:
-        time_range_string = str(stream_request.time_range).replace(" ", "")
-        anno_fname = 'annotations_%s.json' % time_range_string
+        anno_fname = get_annotation_filename(stream_request)
         anno_json = os.path.join(base_path, anno_fname)
         stream_request.annotation_store.dump_json(anno_json)